Choosing dtype float32 or float64

+++ geotiepoints/geointerpolator.py
@@ -86,11 +86,11 @@ class GeoInterpolator(Interpolator):
         # lat_f = tempfile.NamedTemporaryFile(dir='/home/mag/Documents/repos/solab/PySOL/notebooks/POSADA/')
         lat_f = tempfile.NamedTemporaryFile()
         lat_f.name = '/tmp/.lats.npz'
-        lat = memmap(lat_f.name, dtype=float64, mode='w+', shape=shape)
+        lat = memmap(lat_f.name, dtype=newx.dtype, mode='w+', shape=shape)
         # lon_f = tempfile.NamedTemporaryFile(dir='/home/mag/Documents/repos/solab/PySOL/notebooks/POSADA/')
         lon_f = tempfile.NamedTemporaryFile()
         lon_f.name = '/tmp/.lons.npz'
-        lon = memmap(lon_f.name, dtype=float64, mode='w+', shape=shape)
+        lon = memmap(lon_f.name, dtype=newx.dtype, mode='w+', shape=shape)
 
         lon[:] = get_lons_from_cartesian(newx, newy)[:]
         lat[:] = get_lats_from_cartesian(newx, newy, newz)[:]
